have added in a main toolbar but as of yet none of this works. thus the next steps will be enabling all of this.
The plan with this is that the draw operations will be saved to a single 
file for all images. and then they can be exported into a set of PNGs 
(SVG later will see how it goes). the genious thing here is that by 
saving the draw operations undo will still work even if the file is 
loaded new and also it makes autosave much easier.