going to pause the draw svg and draw raster stuff now. The reason being is that the draw operations structure is becoming insanely complex to maintain. there are a lot of arrays there that are taking up empty space.
There are two options here 1) use a union of structs to make it easier 
to map the same pieces of memory but will need a complete rewrite of the 
file read and write. 2) give all the arrays generic names meaning less 
code to rewrite but have to be very careful about naming. will see what 
to do.